Allow to derive the target-index from either the actual source index or the contents of the document
Allow the target-index to be a pattern which is resolved for every document and this way can use either the source-index or values of fields from the current document for defining the resulting target-index. This is useful if you would like to split the index differently or use time-based indexing and still want to reindex in-bulk across multiple indices.